Create battleshipextracredit.py
There has to be some better way to do ship generation. I'm content with only having two ships thus far, since the grid sizes aren't so great. But if I wanted to add a four-space sized ship or a five-space sized ship it seems like it would be a nightmare to implement all the possible combinations for how to add spaces to the ship after initial placement of the first point.